chore: move widgets to cell
- first cell will have widgets.
  Row can no longer have it since its display: contents).
  Plus widget controls are really posesion of the first cell.
