+++ src/components/zen-table-row/zen-table-row.tsx
@@ -1,6 +1,4 @@
 import { h, Component, Host, Prop, Element, Event, EventEmitter, Method, Watch } from '@stencil/core';
-import { applyPrefix } from '../helpers/helpers';
-import { faChevronRight } from '@fortawesome/pro-light-svg-icons';
 
 /**
  * @slot [default] - Content for table cells
@@ -162,31 +160,8 @@ export class ZenTableRow {
   }
 
   render(): HTMLTableRowElement {
-    const ZenCheckBox = applyPrefix('zen-checkbox', this.host);
-    const ZenIcon = applyPrefix('zen-icon', this.host);
     return (
       <Host>
-        {this.showWidgets() && (
-          <div class="widgets">
-            {this.selectable && (
-              <ZenCheckBox
-                indeterminate={this.$indeterminate}
-                class="checkbox"
-                checked={this.selected}
-                onClick={() => this.onSelect()}
-              />
-            )}
-            {this.expandable && (
-              <ZenIcon
-                class="expand-icon"
-                size="sm"
-                padding="lg"
-                icon={faChevronRight}
-                onClick={() => this.onExpand()}
-              />
-            )}
-          </div>
-        )}
         <slot></slot>
       </Host>
     );
